<==== Mercurial data ====>
Branch: paxton

+++ client/src/utils/app_server_file_cache.cpp
@@ -2,27 +2,38 @@
 
 #include <QtCore/QDir>
 #include <QtCore/QFileInfo>
-#include <utils/common/uuid.h>
 #include <QtCore/QTimer>
 #include <QtCore/QStandardPaths>
 
-#include <utils/common/util.h> /* For removeDir. */
-
 #include <api/app_server_connection.h>
 
-#include <ui/workbench/workbench_context.h>
-#include <ui/workbench/workbench_state_manager.h>
+#include <common/common_module.h>
+#include <client/client_message_processor.h>
+
+#include <utils/common/util.h> /* For removeDir. */
+#include <utils/common/uuid.h>
+#include <utils/common/string.h>
 
 QnAppServerFileCache::QnAppServerFileCache(const QString &folderName, QObject *parent) :
     QObject(parent),
-    QnWorkbenchContextAware(parent),
-    m_folderName(folderName),
-    m_workbenchStateDelegate(new QnBasicWorkbenchStateDelegate<QnAppServerFileCache>(this))
+    m_folderName(folderName)
 {
-    connect(this, &QnAppServerFileCache::delayedFileDownloaded,     this,   &QnAppServerFileCache::fileDownloaded,      Qt::QueuedConnection);
-    connect(this, &QnAppServerFileCache::delayedFileUploaded,       this,   &QnAppServerFileCache::fileUploaded,        Qt::QueuedConnection);
-    connect(this, &QnAppServerFileCache::delayedFileDeleted,        this,   &QnAppServerFileCache::fileDeleted,         Qt::QueuedConnection);
-    connect(this, &QnAppServerFileCache::delayedFileListReceived,   this,   &QnAppServerFileCache::fileListReceived,    Qt::QueuedConnection);
+    connect(this, &QnAppServerFileCache::delayedFileDownloaded,     this,   [this](const QString &filename, bool ok) {
+        auto connection = QnAppServerConnectionFactory::getConnection2();
+        emit fileDownloaded(filename, ok && static_cast<bool>(connection));
+    }, Qt::QueuedConnection);
+    connect(this, &QnAppServerFileCache::delayedFileUploaded,     this,   [this](const QString &filename, bool ok) {
+        auto connection = QnAppServerConnectionFactory::getConnection2();
+        emit fileUploaded(filename, ok && static_cast<bool>(connection));
+    }, Qt::QueuedConnection);
+    connect(this, &QnAppServerFileCache::delayedFileDeleted,     this,   [this](const QString &filename, bool ok) {
+        auto connection = QnAppServerConnectionFactory::getConnection2();
+        emit fileDeleted(filename, ok && static_cast<bool>(connection));
+    }, Qt::QueuedConnection);
+    connect(this, &QnAppServerFileCache::delayedFileListReceived,     this,   [this](const QStringList &files, bool ok) {
+        auto connection = QnAppServerConnectionFactory::getConnection2();
+        emit fileListReceived(files, ok && static_cast<bool>(connection));
+    }, Qt::QueuedConnection);
 }
 
 QnAppServerFileCache::~QnAppServerFileCache(){}
@@ -30,12 +41,17 @@ QnAppServerFileCache::~QnAppServerFileCache(){}
 // -------------- Utility methods ----------------
 
 QString QnAppServerFileCache::getFullPath(const QString &filename) const {
+    auto connectionState = qnClientMessageProcessor->connectionState();
+    Q_ASSERT_X(connectionState != QnConnectionState::Disconnected || connectionState == QnConnectionState::Invalid,
+        Q_FUNC_INFO, "Method should be called only when we are know the target system. Current state is " + QnConnectionStateUtils::toString(connectionState).toUtf8());
+
+    /* Avoid empty folder name and collisions with our folders such as 'log'. */
+    QString systemName = L'_' + replaceNonFileNameCharacters(qnCommon->localSystemName(), L'_');
+
     QString path = QStandardPaths::writableLocation(QStandardPaths::DataLocation);
-    QUrl url = QnAppServerConnectionFactory::url();
-    return QDir::toNativeSeparators(QString(QLatin1String("%1/cache/%2_%3/%4/%5"))
+    return QDir::toNativeSeparators(QString(lit("%1/cache/%2/%3/%4"))
                                     .arg(path)
-                                    .arg(url.host(QUrl::FullyEncoded))
-                                    .arg(url.port())
+                                    .arg(systemName)
                                     .arg(m_folderName)
                                     .arg(filename)
                                     );
@@ -52,10 +68,13 @@ QString QnAppServerFileCache::folderName() const {
 
 void QnAppServerFileCache::clearLocalCache() {
     QString path = QStandardPaths::writableLocation(QStandardPaths::DataLocation);
-    QString dir = QDir::toNativeSeparators(QString(QLatin1String("%1/cache/")).arg(path));
+    QString dir = QDir::toNativeSeparators(QString(lit("%1/cache/")).arg(path));
     removeDir(dir);
 }
 
+bool QnAppServerFileCache::isConnectedToServer() const {
+    return QnAppServerConnectionFactory::getConnection2() != NULL;
+}
 
 // -------------- File List loading methods -----
 
@@ -114,7 +133,7 @@ void QnAppServerFileCache::at_fileLoaded( int handle, ec2::ErrorCode errorCode, 
     QString filename = m_loading[handle];
     m_loading.remove(handle);
 
-    if (errorCode != ec2::ErrorCode::ok) {
+    if (errorCode != ec2::ErrorCode::ok || !isConnectedToServer()) {
         emit fileDownloaded(filename, false);
         return;
     }
@@ -170,6 +189,12 @@ void QnAppServerFileCache::at_fileUploaded( int handle, ec2::ErrorCode errorCode
 
     QString filename = m_uploading[handle];
     m_uploading.remove(handle);
+
+    if (!isConnectedToServer()) {
+        emit fileUploaded(filename, false);
+        return;
+    }
+
     const bool ok = errorCode == ec2::ErrorCode::ok;
     if (!ok)
         QFile::remove(getFullPath(filename));
@@ -215,6 +240,9 @@ void QnAppServerFileCache::at_fileDeleted( int handle, ec2::ErrorCode errorCode 
     if (!m_deleting.contains(handle))
         return;
 
+    if (!isConnectedToServer())
+        return;
+
     QString filename = m_deleting[handle];
     m_deleting.remove(handle);
     const bool ok = errorCode == ec2::ErrorCode::ok;
@@ -227,14 +255,4 @@ void QnAppServerFileCache::clear() {
     m_loading.clear();
     m_uploading.clear();
     m_deleting.clear();
-}
-
-bool QnAppServerFileCache::tryClose(bool force) {
-    Q_UNUSED(force);
-    clear();
-    return true;
-}
-
-void QnAppServerFileCache::forcedUpdate() {
-    getFileList();
 }
